requirements.txt
python-dotenv
requests
beautifulsoup4
streamlit
pymongo
textblob
spacy
vaderSentiment
typing